+++ src/server/game/Events/GameEventMgr.cpp
@@ -1131,7 +1131,7 @@ void GameEventMgr::UpdateEventNPCFlags(uint16 event_id)
                 uint32 npcflag = GetNPCFlag(cr);
                 if (const CreatureTemplate* ci = cr->GetCreatureTemplate())
                     npcflag |= ci->npcflag;
-                cr->SetUInt32Value(UNIT_FIELD_NPC_FLAGS, npcflag);
+                cr->SetUInt64Value(UNIT_FIELD_NPC_FLAGS, npcflag);
                 // reset gossip options, since the flag change might have added / removed some
                 //cr->ResetGossipOptions();
             }
