+++ src/server/scripts/Outland/TempestKeep/arcatraz/arcatraz.cpp
@@ -545,7 +545,7 @@ class npc_zerekethvoidzone : public CreatureScript
 
             void Reset() OVERRIDE
             {
-                me->SetUInt32Value(UNIT_FIELD_NPC_FLAGS, 0);
+                me->SetUInt64Value(UNIT_FIELD_NPC_FLAGS, 0);
                 me->setFaction(16);
                 me->SetFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_NOT_SELECTABLE);
